+++ themes/v2/views/dashform/layouts/dashform.php
@@ -7,6 +7,7 @@
 <?php
 $baseUrl = Yii::app()->theme->baseUrl.'/common/';
 $cs = Yii::app()->getClientScript();
+$cs->registerCssFile($baseUrl . 'bootstrap/css/material-bootstrap-wizard.css');
 $cs->registerCssFile($baseUrl . 'bootstrap/css/bootstrap3.min.css');
 $cs->registerCssFile($baseUrl . 'bootstrap/css/material-kit.css');
 $cs->registerCssFile($baseUrl . 'bootstrap/css/bootstrap-editable.css');
@@ -22,6 +23,9 @@ $cs->registerScriptFile($baseUrl . 'theme/scripts/plugins/system/less.min.js');
 $cs->registerScriptFile($baseUrl . 'bootstrap/js/bootstrap3.min.js');
 $cs->registerScriptFile($baseUrl . 'bootstrap/js/bootstrap-editable.min.js');
 $cs->registerScriptFile($baseUrl . 'bootstrap/js/material.min.js');
+$cs->registerScriptFile($baseUrl . 'bootstrap/js/jquery.bootstrap.js');
+$cs->registerScriptFile($baseUrl . 'bootstrap/js/jquery.validate.min.js');
+$cs->registerScriptFile($baseUrl . 'bootstrap/js/material-bootstrap-wizard.js');
 $cs->registerScriptFile($baseUrl . 'introjs/intro.js');
 $cs->registerScriptFile($baseUrl . 'd3js/d3.v4.min.js');
 $cs->registerScriptFile($baseUrl . 'd3js/topojson.v1.min.js');